Moving app ui into the UI directory.
